+++ satella/configuration/schema/__init__.py
@@ -1,0 +1,9 @@
+from .base import CheckerCondition, Descriptor
+from .from_json import descriptor_from_dict
+from .basic import IPv4, Integer, String, Float, Boolean
+from .structs import Union, List, Dict, Caster, create_key
+from .registry import register_custom_descriptor
+
+__all__ = ['CheckerCondition', 'Descriptor', 'descriptor_from_dict', 'IPv4', 'Integer',
+           'String', 'Float', 'Boolean', 'Union', 'List', 'Dict', 'Caster', 'register_custom_descriptor',
+           'create_key']